Add pool.py module
